后台各类型总数

+++ time-common/src/main/java/com/sky/mapper/LogOperInfoMapper.java
@@ -13,5 +13,5 @@ import java.util.List;
 public interface LogOperInfoMapper extends BaseMapper<LogOperInfo> {
     List<DataDTO> getDataDays();
 
-    Long getLoginCount();
+    int getLoginCount();
 }
